# docs/content/api/specs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Gasper Kaze API",
    "description": "Handles authentication, creation/management of applications, databases, users and also provides a superuser API. <br><br> Only a superuser can avail the superuser API. A superuser can **grant/revoke** superuser privileges to other users. A default  superuser is created every time a Gasper instance is launched whose credentials are defined in the `admin` section of `config.toml`, the main configuration file. A sample configuration file is available [here](https://github.com/sdslabs/gasper/blob/develop/config.sample.toml#L37).<br><br>\n**Note:-** Normally the applications and databases can only be managed by their owners but the superuser can bypass that check.<br><br>\n**PS:-** If you want to programmatically generate a client for this API, you can find the corresponding OpenAPI specifications [here](https://github.com/sdslabs/gasper/tree/develop/docs/content/api/specs). We recommend using [OpenAPI-Generator](https://openapi-generator.tech/) for generating clients.",
    "contact": {
      "name": "SDSLabs",
      "url": "https://chat.sdslabs.co",
      "email": "contact@sdslabs.co.in"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/sdslabs/gasper/blob/develop/LICENSE.md"
    },
    "version": "1.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "components": {
    "schemas": {
      "Context": {
        "type": "object",
        "required": [
          "index"
        ],
        "properties": {
          "index": {
            "type": "string",
            "description": "Entrypoint for an application",
            "example": "index.js"
          },
          "port": {
            "type": "integer",
            "format": "int64",
            "description": "Port on which the application runs",
            "example": 3000,
            "default": 80
          },
          "rc_file": {
            "type": "boolean",
            "description": "Is a Run Commands file present in the application's git repository's root directory?",
            "default": false
          },
          "build": {
            "type": "array",
            "description": "Commands for installing application dependencies",
            "items": {
              "type": "string"
            },
            "example": [
              "npm install",
              "grunt build"
            ]
          },
          "run": {
            "type": "array",
            "description": "Commands for running the application",
            "items": {
              "type": "string"
            },
            "example": [
              "node index.js"
            ]
          }
        }
      },
      "Git": {
        "type": "object",
        "required": [
          "repo_url"
        ],
        "properties": {
          "repo_url": {
            "type": "string",
            "description": "Git Repository URL of the application",
            "example": "https://github.com/sdslabs/gasper"
          },
          "access_token": {
            "type": "string",
            "description": "Access token for cloning the application's git repository"
          },
          "branch": {
            "type": "string",
            "description": "Specific branch to clone",
            "default": "master",
            "example": "master"
          }
        }
      },
      "Application": {
        "type": "object",
        "required": [
          "name",
          "password",
          "git",
          "context"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the application"
          },
          "password": {
            "type": "string",
            "description": "Password required for SSH access to the application's docker container"
          },
          "git": {
            "$ref": "#/components/schemas/Git"
          },
          "context": {
            "$ref": "#/components/schemas/Context"
          },
          "env": {
            "type": "object",
            "description": "Environment variables required by the application",
            "additionalProperties": {
              "type": "string"
            }
          },
          "resources": {
            "type": "object",
            "description": "Resources required by the application",
            "properties": {
              "cpu": {
                "type": "number",
                "format": "float",
                "example": 0.25,
                "description": "Number of virtual CPUs"
              },
              "memory": {
                "type": "number",
                "format": "float",
                "example": 0.5,
                "description": "Memory in GigaBytes (GB)"
              }
            }
          }
        }
      },
      "CreatedApplication": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the application"
          },
          "password": {
            "type": "string",
            "description": "Password required for SSH access to the application's docker container"
          },
          "git": {
            "$ref": "#/components/schemas/Git"
          },
          "context": {
            "$ref": "#/components/schemas/Context"
          },
          "env": {
            "type": "object",
            "description": "Environment variables required by the application",
            "additionalProperties": {
              "type": "string"
            }
          },
          "resources": {
            "type": "object",
            "description": "Resources required by the application",
            "properties": {
              "cpu": {
                "type": "number",
                "format": "float",
                "example": 0.25,
                "description": "Number of virtual CPUs"
              },
              "memory": {
                "type": "number",
                "format": "float",
                "example": 0.5,
                "description": "Memory in GigaBytes (GB)"
              }
            }
          },
          "container_id": {
            "type": "string",
            "description": "ID of the application's docker container"
          },
          "container_port": {
            "type": "integer",
            "format": "int64",
            "example": 35000,
            "description": "Port assigned by the node to the application's docker container"
          },
          "docker_image": {
            "type": "string",
            "description": "Docker image used in building the application's container",
            "example": "sdsws/node:2.0"
          },
          "app_url": {
            "type": "string",
            "description": "The domain name of the application (DNS entry is managed by GenDNS 💡)",
            "example": "alphadose.app.sdslabs.co"
          },
          "host_ip": {
            "type": "string",
            "example": "192.168.208.206",
            "description": "IPv4 address of the node"
          },
          "name_servers": {
            "type": "array",
            "description": "The DNS NameServers used by the application's docker container",
            "items": {
              "type": "string"
            },
            "example": [
              "8.8.8.8",
              "8.8.4.4"
            ]
          },
          "instance_type": {
            "type": "string",
            "description": "The kind of instance this application belongs to",
            "example": "app"
          },
          "language": {
            "type": "string",
            "description": "The programming language in which the application is written",
            "example": "nodejs",
            "enum": [
              "php",
              "nodejs",
              "static",
              "python2",
              "python3",
              "golang",
              "ruby"
            ]
          },
          "owner": {
            "type": "string",
            "description": "Owner of the application",
            "example": "anish.mukherjee1996@gmail.com"
          },
          "ssh_cmd": {
            "type": "string",
            "description": "Command to SSH into the application's docker container",
            "example": "ssh -p 2222 alphadose@192.168.208.206"
          },
          "_id": {
            "type": "string",
            "description": "Application's identifier assigned by MongoDB"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/apps/{language}": {
      "post": {
        "tags": [
          "apps"
        ],
        "summary": "Create an application",
        "operationId": "createApp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Application"
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "Bearer Token Authentication",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
            }
          },
          {
            "in": "path",
            "name": "language",
            "required": true,
            "description": "The programming language in which the application is written",
            "schema": {
              "type": "string",
              "enum": [
                "static",
                "php",
                "nodejs",
                "python2",
                "python3",
                "golang",
                "ruby"
              ]
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "name": {
                      "type": "string",
                      "description": "Name of the application"
                    },
                    "password": {
                      "type": "string",
                      "description": "Password required for SSH access to the application's docker container"
                    },
                    "git": {
                      "$ref": "#/components/schemas/Git"
                    },
                    "context": {
                      "$ref": "#/components/schemas/Context"
                    },
                    "env": {
                      "type": "object",
                      "description": "Environment variables required by the application",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "resources": {
                      "type": "object",
                      "description": "Resources required by the application",
                      "properties": {
                        "cpu": {
                          "type": "number",
                          "format": "float",
                          "example": 0.25,
                          "description": "Number of virtual CPUs"
                        },
                        "memory": {
                          "type": "number",
                          "format": "float",
                          "example": 0.5,
                          "description": "Memory in GigaBytes (GB)"
                        }
                      }
                    },
                    "container_id": {
                      "type": "string",
                      "description": "ID of the application's docker container"
                    },
                    "container_port": {
                      "type": "integer",
                      "format": "int64",
                      "example": 35000,
                      "description": "Port assigned by the node to the application's docker container"
                    },
                    "docker_image": {
                      "type": "string",
                      "description": "Docker image used in building the application's container",
                      "example": "sdsws/node:2.0"
                    },
                    "app_url": {
                      "type": "string",
                      "description": "The domain name of the application (DNS entry is managed by GenDNS 💡)",
                      "example": "alphadose.app.sdslabs.co"
                    },
                    "host_ip": {
                      "type": "string",
                      "example": "192.168.208.206",
                      "description": "IPv4 address of the node"
                    },
                    "name_servers": {
                      "type": "array",
                      "description": "The DNS NameServers used by the application's docker container",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "8.8.8.8",
                        "8.8.4.4"
                      ]
                    },
                    "instance_type": {
                      "type": "string",
                      "description": "The kind of instance this application belongs to",
                      "example": "app"
                    },
                    "language": {
                      "type": "string",
                      "description": "The programming language in which the application is written",
                      "example": "nodejs",
                      "enum": [
                        "php",
                        "nodejs",
                        "static",
                        "python2",
                        "python3",
                        "golang",
                        "ruby"
                      ]
                    },
                    "owner": {
                      "type": "string",
                      "description": "Owner of the application",
                      "example": "anish.mukherjee1996@gmail.com"
                    },
                    "ssh_cmd": {
                      "type": "string",
                      "description": "Command to SSH into the application's docker container",
                      "example": "ssh -p 2222 alphadose@192.168.208.206"
                    },
                    "_id": {
                      "type": "string",
                      "description": "Application's identifier assigned by MongoDB"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Token is expired"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "INTERNAL_SERVER_ERROR"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/apps": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "Fetch all applications owned by a user",
        "operationId": "fetchAppsByUser",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "description": "Bearer Token Authentication",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CreatedApplication"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Token is expired"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "INTERNAL_SERVER_ERROR"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}